fix(temporal): remove schedule definition import cycle
Why: The final Buildkite verification found that splitting the schedule registry introduced an eager import cycle, which the architecture gate correctly rejected. What: move ScheduleDefinition and CatchupWindow into a neutral schedule-types module and update both registry consumers to depend on it, preserving the schedule list and runtime behavior. Verification: bun x eslint src/schedules/schedule-types.ts src/schedules/schedule-definitions.ts src/schedules/schedule-definitions-early.ts src/schedules/register-schedules.ts; bun x tsc --noEmit --pretty false -p tsconfig.json; bun test src/schedules/register-schedules.test.ts; bunx lefthook run pre-commit. Buildkite rerun pending.
